+++ src/Farmer/ArmBuilder.fs
@@ -4,15 +4,27 @@ module Farmer.ArmBuilder
 open Farmer.CoreTypes
 open Farmer
 
-module Helpers =
+module Resource =
     /// Creates a unique IArmResource from an arbitrary object.
-    let toArmResource armObject =
+    let ofObj armObject =
         { new IArmResource with
              member _.ResourceName = ResourceName (System.Guid.NewGuid().ToString())
              member _.JsonModel = armObject }
 
-    /// Creates a Builder that can be added to arm { } expressions.
-    let asBuilder quickBuilder  =
+    /// Creates a unique IArmResource from a JSON string containing the output you want.
+    let ofJson json = json |> Newtonsoft.Json.Linq.JObject.Parse |> ofObj
+
+module Json =
+    /// Creates a unique IArmResource from a JSON string containing the output you want.
+    let toIArmResource = Resource.ofJson
+
+module Subscription =
+    /// Gets an ARM expression pointing to the tenant id of the current subscription.
+    let TenantId = ArmExpression "subscription().tenantid"
+
+module Builder =
+    /// Quickly creates a Builder that can be added to arm { } expressions.
+    let fromFunction quickBuilder =
         let output : Builder = fun location _ -> quickBuilder location
         output